Update weekplan_screen.dart
fontSize for now set to 30. A different issue needs to be created for the same approach of text size throughout the weekplanner app. It is implemented differently and many places hardcoded. Also are the minimum text requirement neccessary to be so high or to be there at all (see design guide)? It will be weird for small widgets.