Update weekplan_screen.dart  (#313)
* Update weekplan_screen.dart

Changed to use AutoSizeText() instead of Text() for the Card widget in in weekplan_screen.dart

* Update weekplan_screen.dart

Assigned an AutoSizeGroup to have same AutoSizeText for all Card titles as well as setting padding to zero (in correspondence with PO group). As suggested by @giano574

* Update weekplan_screen.dart

Fixed Azure Pipelines Linter error with an unneccesary new keyword used.

* Update weekplan_screen.dart

Removed line specifying minFontSize, this was not originally there and was rather arbitrary set to 8.

* Update weekplan_screen.dart

fontSize for now set to 30. A different issue needs to be created for the same approach of text size throughout the weekplanner app. It is implemented differently and many places hardcoded. Also are the minimum text requirement neccessary to be so high or to be there at all (see design guide)? It will be weird for small widgets.