Changed pictogram box changes
The Issue was that some text was cut off when it spanned multiple lines, as the box size was static. this has been solved by allowing the pictogram box to be as high as the elements within and set the max amount of lines for the text to 2.

this made smaller words split into 2 lines, which was further solved by only allowing the text to be 1 line of only 1 word was present in the string.