Update package link (#554)
This organization has been renamed, and so the name is up for grabs. Currently, it still works because Github redirects to the renamed organization page, but if the old name is registered, the redirection will stop, and thus many tests will break. This update will ensure that the link won't break.