develop to feature/434 (#566)
* Update package link (#554)

This organization has been renamed, and so the name is up for grabs. Currently, it still works because Github redirects to the renamed organization page, but if the old name is registered, the redirection will stop, and thus many tests will break. This update will ensure that the link won't break.

* Feature/562 Updates to unit tests (#564)

* Updated unit tests

* Update weekplans_bloc_test.dart

* FIxes for linter

Co-authored-by: Bogi Napoleon Wennerstrøm <bogi.wennerstrom@gmail.com>
Co-authored-by: Jakob Precht <sporvogn@live.com>