Refactoring of font sizes (#599)
* Refactoring of font sizes

Font sizes now have a document where they are defined. There was used a font size of 25 twice in upload_image_from_phone_screen which has been set to 24 so it's the same as the other places. Likewise the font size was set to 18 in weekplan_selector_screen and it's been set to 20 as in line with the rest of the app.

* Update font_size.dart

* Update weekplan_selector_screen.dart

* Update font_size.dart