feature/696: Inconsistent buttons (#701)
* Inconsistent buttons

Only partly solved the issue.
The buttons now correctly remove the timer and add choiceboard options when marking an activity as completed as well as cancelled. Furthermore an isEnabled has been added to the cancel button, so it becomes disables when the state is completed.
The buttons however only become disabled, when you reenter the activity. This is due to the fact the other button does not check for a new state, when the first button is pressed and as such does not get the new state information, when an acitivty is completed or cancelled. There is not enough time to fix this before release, and as such only the most system breaking part of the bug has been fixed, as it would do some funky things, when trying to make a choiceboard on a completed activity.

* Update show_activity_screen.dart

Fixed missing Add Choiceboard on new activities