Removed orientation, colour theme and timer control along with adding user settings availability (#857)
In terms of nessecary settings the removal of the following from the settings tab due to it being deemed not as nessecary as other aspects:
- Locking the orientation of the app.
- Changing the colour theme of a week.
- Locking the timer control.
These may be brought up again later, but for now they have been hidden in the app.

Also, the addition that a Guardian can select whether a citizen can see their own settings and change them.